update exported types
