+++ src/index.ts
@@ -1,20 +1,5 @@
-import { FSInvalidArgumentError } from './errors/invalidArgument';
-import { FullStoryClient, FullStoryImpl } from './fullstory';
-import { FullStoryOptions } from './http';
-
-export function init(opts: FullStoryOptions): FullStoryClient {
-    if (!opts.apiKey) {
-        throw new FSInvalidArgumentError('The apiKey is required in opts.');
-    }
-    const apiKey = opts.apiKey.indexOf(' ') < 0 ? `Basic ${opts.apiKey}` : opts.apiKey;
-    return new FullStoryImpl({ ...opts, apiKey });
-}
-
-export { BatchJobOptions } from './batch';
-export { FSError, FSErrorName, isFSError } from './errors';
-export { BatchEventsJob, Events } from './events';
-export { FullStoryClient } from './fullstory';
-export { FullStoryOptions } from './http';
-export { BatchUsersJob, Users } from './users';
+export * from './events';
+export * from './fullstory';
+export * from './users';
 export * from '@api/index';
 export * from '@model/index';
